Github actions cache clients expect 204 when there is no entry
While the official github actions cache client is forgiving, others like
opendal are not. Opendal is used by sccache's ghac backend, and
currently sccache crashes when it encounters a missing entry, because it
does not do the check for archive location being present; instead it
checks for the 204 status code.