[WARP-571] feat: windows telemetry support (#38)
* add dummy.exe

* upload artifact

* Renormalize files for LFS

* add windows otel

* [skip ci] add readme for otel windows

* [skip ci] downlaod agentd from s3

* handle 'unexpected end of JSON input' for settings file

* add log for retry

* push to cloudflare r2 instead

* skip configure (read from env)

* use rclone for upload

aws s3 cli is incomptible.

* [skip ci] use r2 for configuring agents

* [skip ci] debug log

* use public dev url to pull the data from r2

* fix name

no bucket name when using dev url

* fix: unexpected JSON end

handle unexpected JSON end for settings file. We retry in this case.

* feat: increate the buffer size to 1MB

* try telemetryd to be started with auto delayed

* fix: delay start for telemetryd

* fix: path escape on windows

* use single quotes for string

* change receivers in case of windows

windows formatting is different which doesn't seem to render properly
in our log files. Change to windowseventlog/system as receiver on
windows.

* log events in console for debug

* write to a json file

* enable per minute metrics

* add services and windowseventlog

* remove restarter logs

* no log running

* skip re-logging in case of running

* remove batch size

* try diff source for metrics

* update batch size to 100

* try watch on dir level for win

* add debug logs

* log watchlist

* try on multiple events

* use append

* split logs and metrics collection files

* rotation not supported with append

turning off rotation

* feat: port for listening for otel events

Moves logic from fileread.

* update otel config file to use metrics as exporter

* vendor sync

* fix: upload logs and metrics

Chunk size is decreased to 100 lines. Since flush logic is not
working currently. Minimizing lines will reduce dataloss when
system is going to sleeping.

Seperation of metrics and logs are added.

Otel config formatting is fixed. It was sending gzipped data instead
of raw json.

* fix: incorrect file uploads for metrics and logs

* use r2 to download warp packages instead

Fallsback to a default tag version which is available. Default tag
is manually verified to be available in r2.

* fix: incorrect fallback tag

* fix: repeated lines in upload

When uploading the events to s3. The uploaded events aren't being
removed, this leads to a lot of duplicate data that we don't need.
We change the logic here to clean the buffer after sending the
data to s3 uploader.

* add cpu if os is not windows in hostmetrics

* simply code

* init buffer if not done

* fix: telemetry issues

* add logging middleware for debug

* try sending faster

* Add log level support

Since logs are contantly need for debugging, etc. A field to control the
log level of the warpbuild services has been introduced. This can
be passed as a flag when booting up warpbuild-agentd or warpbuild-telemetryd.
Defaults to info level logs.

* remove debug level logs from otel

* move logic to otel collector

* revert readme

* revert syslog hardcoding

* syslog hardcoding removed

* revert .gitignore

* fix formatting

* use packages.warbuild.com for url